util: asynctestcase: AsyncExitStackTestCase
