+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `AsyncExitStackTestCase` which instantiates and enters async and non-async
+  `contextlib` exit stacks. Provides temporary file creation.
+### Changed
+- CLI tests and integration tests derive from `AsyncExitStackTestCase`
 ### Fixed
 - Correctly identify when functions decorated with `op` use `self` to reference
   the `OperationImplementationContext`.
