+++ dffml/util/asynctestcase.py
@@ -4,10 +4,15 @@
 Adds support for test cases which need to be run in an event loop.
 """
 import os
+import random
+import pathlib
 import asyncio
 import inspect
 import logging
 import unittest
+import tempfile
+import contextlib
+from typing import Optional
 
 
 class AsyncTestCase(unittest.TestCase):
@@ -61,3 +66,34 @@ class AsyncTestCase(unittest.TestCase):
             ):
                 setattr(self, name, self.async_wrapper(method))
         return super().run(result=result)
+
+
+@contextlib.contextmanager
+def non_existant_tempfile():
+    """
+    Yield the filename of a non-existant file within a temporary directory
+    """
+    with tempfile.TemporaryDirectory() as testdir:
+        yield os.path.join(testdir, str(random.random()))
+
+
+class AsyncExitStackTestCase(AsyncTestCase):
+    async def setUp(self):
+        super().setUp()
+        self._stack = contextlib.ExitStack().__enter__()
+        self._astack = await contextlib.AsyncExitStack().__aenter__()
+
+    async def tearDown(self):
+        super().tearDown()
+        self._stack.__exit__(None, None, None)
+        await self._astack.__aexit__(None, None, None)
+
+    def mktempfile(
+        self, suffix: Optional[str] = None, text: Optional[str] = None
+    ):
+        filename = self._stack.enter_context(non_existant_tempfile())
+        if suffix:
+            filename = filename + suffix
+        if text:
+            pathlib.Path(filename).write_text(inspect.cleandoc(text) + "\n")
+        return filename
